Remove duplication from AMD/r.js configuration
By setting the `baseUrl` to be the `bower_components` directory, limit
the duplication between the general AMD configuration (used on its own
during development) and the "r.js optimizer" configuration (used in
conjunction with the AMD configuration when building the project). In
this way, only a static number of paths (namely `activities/`,
`components/`, and `scripts/`) need to be overridden at build
time--vendor scripts can use identical pathing information across
environments.

Furthermore, this approach makes AMD loading function in a way that is
more consistent with Node.js's implementation of CommonJS (where
non-relative paths are resolve to vendor modules).