Add ReportResource factory and Template rendering streams.
ReportResource stream loads the data stream collected for a room and
then transforms it by a given option.

TemplateReportStream and PrepareStream helper will take an aggregate
data object and produce a rendered template. PrepareStream takes the
aggregate object, storing it in a new object at a report field, and
then further enriches that object before writing it.
TemplateReportStream writes its rendered templated to a render field of
the objects it reads.