Update module dependency
The latest release of the Cloak library supports AMD, so the AMD "shim
configuration" can be removed. Because Cloak now depends on a module
with the ID "socket.io-client" and because this path is more accurate
than the current alias ("socket.io"), the CEE codebase should be updated
to comply.