PPF: Improve drag-and-drop
In order to determine if a given "drop" operation is valid (meaning the
element being dropped should retain its current position) or invalid
(meaning the element being dropped should revert to its original
location), the slop value must be read from the element being dragged.
Previously, the heuristic used to find the element being dragged would
not work if the drag event began with a `target` of a *child* of the
draggable element (i.e. the text label describing the slope value).

Use the context of the jQuery.pep `revertIf` method (which was not
previously defined [1]) to consistently reference the draggable element
(regardless of which element began the drag operation).

[1] This functionality was included in the 0.5.14 release of the jQuery
.pep library:

    > Set jQuery object as context of `revertIf` call
    >
    > This allows custom implementations of `revertIf` to reference the
    > current selection without relying on an externally-scoped
    > reference.

    https://github.com/briangonzalez/jquery.pep.js/commit/db160f02764b104e17e480ea1605191b22160dfc