Pizza Prod: Defer marking players as "ready"
Introduce a new boolean attribute for players named `isReady`.
Initialize this attribute to `false`, and set to `true` when the player
dismisses the modal dialog containing the game instructions.

Update the server logic to wait for the requisite number of players to
be "ready" as defined above (instead of simply being present).

This change ensures that games do not begin until all players have had
an opportunity to read the instructions.