Print port on command line
The server's port number is not always explicitly specified (it may be
assigned dynamically by the operating system or it may be set via
wrapper scripts). Print the port number to standard output in order to
make the port number easier to discover when the server is started from
the command line.