Update third-party dependency
The required version of the `node-http-proxy` package specifies its
single dependency (on the `eventemitter3` package) using the "*" version
pattern. Backward-incompatable changes to `eventemitter3` have since
been published, which in turn broke all versions of `node-http-proxy`
using the permissive version pattern.

Update to the latest version of `node-http-proxy`, which explicitly
specifies the versions of its dependencies. This fixes the immediate
problem and avoids future regressions.