Pizza Prod: Fix bug in player navigation
Clients are disallowed from navigating between workstations as long as
they are in possession of a pizza. Ensure that all pizzas ("active" or
"inactive") are unassigned at the termination of each round so that
players who are in possession of a pizza as a round ends are able to
navigate immediately at the onset of the following round.