Pizza Prod: Correct message semantics
Due to the way Backbone.js implements methods like `save`, models
without `id` attributes are considered "new" and are therefore
synchronized with "create" messages. Because the game state is created
when the game begins and lives throughout the session, calls to `save`
should actually trigger "update" messages.

Due to the simplicity of the current implementation, this distinction
has no practical effect on program operation. Correct the message
semantics in order to avoid issues with future extensions.