Pizza Prod: Synchronize endgame report
Ensure that last-minute changes to the game's state are reflected in the
endgame report for all clients.

Extensions to the client-server protocol (e.g. a formal "endgame"
message) could preclude the need for such dynamism, but that
verification step would also introduce a delay between game completion
and report rendering, decreasing perceived responsiveness.

Because the race condition under consideration is relatively rare, the
(user-perceived) advantages of the dynamic approach outweigh the
(largely technical) drawbacks of its less formal protocol.