Pizza Prod: Fix "queue pizza" placement issues
Ensure that when "queue pizzas" are removed from the queue, their
position on the space is reserved (avoiding visual shifting of the
remaining pizzas).

In some circumstances (e.g. when one player takes and releases a pizza
while another player is dragging it), "queue pizzas" would not revert to
their correct location. This change also corrects the placement in that
case.